+++ packages/react-native-the-sheet/src/core/private/hooks/use-to-shared-value.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react'
 import { isSharedValue, useSharedValue } from 'react-native-reanimated'
-import type { AnimatedProp } from '../types'
+import type { AnimatedProp } from '../../types'
 
 export function useToSharedValue<T>(value: AnimatedProp<T>) {
   const shared = useSharedValue<T>(